fixed single card
